fix: prepare for ep review

+++ src/main/java/io/supertokens/storage/postgresql/queries/EmailVerificationQueries.java
@@ -20,7 +20,6 @@ import io.supertokens.pluginInterface.RowMapper;
 import io.supertokens.pluginInterface.emailverification.EmailVerificationTokenInfo;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
-import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.storage.postgresql.Start;
 import io.supertokens.storage.postgresql.config.Config;
 import io.supertokens.storage.postgresql.utils.Utils;
@@ -43,15 +42,9 @@ public class EmailVerificationQueries {
         String emailVerificationTable = Config.getConfig(start).getEmailVerificationTable();
         // @formatter:off
         return "CREATE TABLE IF NOT EXISTS " + emailVerificationTable + " ("
-                + "app_id VARCHAR(64) DEFAULT 'public',"
                 + "user_id VARCHAR(128) NOT NULL,"
                 + "email VARCHAR(256) NOT NULL,"
-                + "CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTable, null, "pkey")
-                + " PRIMARY KEY (app_id, user_id, email),"
-                + "CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTable, "app_id", "fkey")
-                + " FOREIGN KEY(app_id)"
-                + " REFERENCES " + Config.getConfig(start).getAppsTable() +  "(app_id) ON DELETE CASCADE"
-                + ");";
+                + "CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTable, null, "pkey") + " PRIMARY KEY (user_id, email));";
         // @formatter:on
     }
 
@@ -60,13 +53,12 @@ public class EmailVerificationQueries {
         String emailVerificationTokensTable = Config.getConfig(start).getEmailVerificationTokensTable();
         // @formatter:off
         return "CREATE TABLE IF NOT EXISTS " + emailVerificationTokensTable + " ("
-                + "app_id VARCHAR(64) DEFAULT 'public',"
                 + "user_id VARCHAR(128) NOT NULL,"
                 + "email VARCHAR(256) NOT NULL,"
                 + "token VARCHAR(128) NOT NULL CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTokensTable, "token", "key") + " UNIQUE,"
                 + "token_expiry BIGINT NOT NULL,"
-                + "CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTokensTable, null, "pkey")
-                + " PRIMARY KEY (app_id, user_id, email, token))";
+                + "CONSTRAINT " + Utils.getConstraintName(schema, emailVerificationTokensTable, null, "pkey") +
+                " PRIMARY KEY (user_id, email, token))";
         // @formatter:on
     }
 
@@ -81,53 +73,44 @@ public class EmailVerificationQueries {
         update(start, QUERY, pst -> pst.setLong(1, currentTimeMillis()));
     }
 
-    public static void updateUsersIsEmailVerified_Transaction(Start start, Connection con, AppIdentifier appIdentifier,
-                                                              String userId, String email,
-                                                              boolean isEmailVerified) throws SQLException, StorageQueryException {
+    public static void updateUsersIsEmailVerified_Transaction(Start start, Connection con, String userId, String email,
+            boolean isEmailVerified) throws SQLException, StorageQueryException {
 
         if (isEmailVerified) {
             String QUERY = "INSERT INTO " + getConfig(start).getEmailVerificationTable()
-                    + "(app_id, user_id, email) VALUES(?, ?, ?)";
+                    + "(user_id, email) VALUES(?, ?)";
 
             update(con, QUERY, pst -> {
-                pst.setString(1, appIdentifier.getAppId());
-                pst.setString(2, userId);
-                pst.setString(3, email);
+                pst.setString(1, userId);
+                pst.setString(2, email);
             });
         } else {
             String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTable()
-                    + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                    + " WHERE user_id = ? AND email = ?";
 
             update(con, QUERY, pst -> {
-                pst.setString(1, appIdentifier.getAppId());
-                pst.setString(2, userId);
-                pst.setString(3, email);
+                pst.setString(1, userId);
+                pst.setString(2, email);
             });
         }
     }
 
-    public static void deleteAllEmailVerificationTokensForUser_Transaction(Start start, Connection con,
-                                                                           AppIdentifier appIdentifier, String userId,
-                                                                           String email) throws SQLException, StorageQueryException {
+    public static void deleteAllEmailVerificationTokensForUser_Transaction(Start start, Connection con, String userId,
+            String email) throws SQLException, StorageQueryException {
         String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTokensTable()
-                + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                + " WHERE user_id = ? AND email = ?";
 
         update(con, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         });
     }
 
-    public static EmailVerificationTokenInfo getEmailVerificationTokenInfo(Start start, AppIdentifier appIdentifier,
-                                                                           String token)
+    public static EmailVerificationTokenInfo getEmailVerificationTokenInfo(Start start, String token)
             throws SQLException, StorageQueryException {
         String QUERY = "SELECT user_id, token, token_expiry, email FROM "
-                + getConfig(start).getEmailVerificationTokensTable() + " WHERE app_id = ? AND token = ?";
-        return execute(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, token);
-        }, result -> {
+                + getConfig(start).getEmailVerificationTokensTable() + " WHERE token = ?";
+        return execute(start, QUERY, pst -> pst.setString(1, token), result -> {
             if (result.next()) {
                 return EmailVerificationTokenInfoRowMapper.getInstance().mapOrThrow(result);
             }
@@ -135,32 +118,28 @@ public class EmailVerificationQueries {
         });
     }
 
-    public static void addEmailVerificationToken(Start start, AppIdentifier appIdentifier, String userId, String tokenHash, long expiry,
-                                                 String email) throws SQLException, StorageQueryException {
+    public static void addEmailVerificationToken(Start start, String userId, String tokenHash, long expiry,
+            String email) throws SQLException, StorageQueryException {
         String QUERY = "INSERT INTO " + getConfig(start).getEmailVerificationTokensTable()
-                + "(app_id, user_id, token, token_expiry, email)" + " VALUES(?, ?, ?, ?, ?)";
+                + "(user_id, token, token_expiry, email)" + " VALUES(?, ?, ?, ?)";
 
         update(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, tokenHash);
-            pst.setLong(4, expiry);
-            pst.setString(5, email);
+            pst.setString(1, userId);
+            pst.setString(2, tokenHash);
+            pst.setLong(3, expiry);
+            pst.setString(4, email);
         });
     }
 
     public static EmailVerificationTokenInfo[] getAllEmailVerificationTokenInfoForUser_Transaction(Start start,
-                                                                                                   Connection con,
-                                                                                                   AppIdentifier appIdentifier,
-                                                                                                   String userId, String email) throws SQLException, StorageQueryException {
+            Connection con, String userId, String email) throws SQLException, StorageQueryException {
 
         String QUERY = "SELECT user_id, token, token_expiry, email FROM "
-                + getConfig(start).getEmailVerificationTokensTable() + " WHERE app_id = ? AND user_id = ? AND email = ? FOR UPDATE";
+                + getConfig(start).getEmailVerificationTokensTable() + " WHERE user_id = ? AND email = ? FOR UPDATE";
 
         return execute(con, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         }, result -> {
             List<EmailVerificationTokenInfo> temp = new ArrayList<>();
             while (result.next()) {
@@ -174,17 +153,14 @@ public class EmailVerificationQueries {
         });
     }
 
-    public static EmailVerificationTokenInfo[] getAllEmailVerificationTokenInfoForUser(Start start,
-                                                                                       AppIdentifier appIdentifier,
-                                                                                       String userId,
-                                                                                       String email) throws SQLException, StorageQueryException {
+    public static EmailVerificationTokenInfo[] getAllEmailVerificationTokenInfoForUser(Start start, String userId,
+            String email) throws SQLException, StorageQueryException {
         String QUERY = "SELECT user_id, token, token_expiry, email FROM "
-                + getConfig(start).getEmailVerificationTokensTable() + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                + getConfig(start).getEmailVerificationTokensTable() + " WHERE user_id = ? AND email = ?";
 
         return execute(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         }, result -> {
             List<EmailVerificationTokenInfo> temp = new ArrayList<>();
             while (result.next()) {
@@ -198,40 +174,32 @@ public class EmailVerificationQueries {
         });
     }
 
-    public static boolean isEmailVerified(Start start, AppIdentifier appIdentifier, String userId, String email)
+    public static boolean isEmailVerified(Start start, String userId, String email)
             throws SQLException, StorageQueryException {
         String QUERY = "SELECT * FROM " + getConfig(start).getEmailVerificationTable()
-                + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                + " WHERE user_id = ? AND email = ?";
 
         return execute(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         }, result -> result.next());
     }
 
-    public static void deleteUserInfo(Start start, AppIdentifier appIdentifier, String userId)
+    public static void deleteUserInfo(Start start, String userId)
             throws StorageQueryException, StorageTransactionLogicException {
         start.startTransaction(con -> {
             Connection sqlCon = (Connection) con.getConnection();
             try {
                 {
-                    String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTable()
-                            + " WHERE app_id = ? AND user_id = ?";
-                    update(sqlCon, QUERY, pst -> {
-                        pst.setString(1, appIdentifier.getAppId());
-                        pst.setString(2, userId);
-                    });
+                    String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTable() + " WHERE user_id = ?";
+                    update(sqlCon, QUERY, pst -> pst.setString(1, userId));
                 }
 
                 {
                     String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTokensTable()
-                            + " WHERE app_id = ? AND user_id = ?";
+                            + " WHERE user_id = ?";
 
-                    update(sqlCon, QUERY, pst -> {
-                        pst.setString(1, appIdentifier.getAppId());
-                        pst.setString(2, userId);
-                    });
+                    update(sqlCon, QUERY, pst -> pst.setString(1, userId));
                 }
 
                 sqlCon.commit();
@@ -242,37 +210,34 @@ public class EmailVerificationQueries {
         });
     }
 
-    public static void unverifyEmail(Start start, AppIdentifier appIdentifier, String userId, String email)
+    public static void unverifyEmail(Start start, String userId, String email)
             throws SQLException, StorageQueryException {
         String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTable()
-                + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                + " WHERE user_id = ? AND email = ?";
 
         update(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         });
     }
 
-    public static void revokeAllTokens(Start start, AppIdentifier appIdentifier, String userId, String email)
+    public static void revokeAllTokens(Start start, String userId, String email)
             throws SQLException, StorageQueryException {
         String QUERY = "DELETE FROM " + getConfig(start).getEmailVerificationTokensTable()
-                + " WHERE app_id = ? AND user_id = ? AND email = ?";
+                + " WHERE user_id = ? AND email = ?";
 
         update(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
-            pst.setString(3, email);
+            pst.setString(1, userId);
+            pst.setString(2, email);
         });
     }
 
-    public static boolean isUserIdBeingUsedForEmailVerification(Start start, AppIdentifier appIdentifier, String userId)
+    public static boolean isUserIdBeingUsedForEmailVerification(Start start, String userId)
             throws SQLException, StorageQueryException {
-        String QUERY = "SELECT * FROM " + getConfig(start).getEmailVerificationTokensTable() + " WHERE app_id = ? AND user_id = ?";
+        String QUERY = "SELECT * FROM " + getConfig(start).getEmailVerificationTokensTable() + " WHERE user_id = ?";
 
         return execute(start, QUERY, pst -> {
-            pst.setString(1, appIdentifier.getAppId());
-            pst.setString(2, userId);
+            pst.setString(1, userId);
         }, ResultSet::next);
     }
